Further renaming Refactor and fix bug with localstorage
