+++ src/router/index.js
@@ -2,9 +2,9 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import TopListView from '@/views/TopListView'
 import NewListView from '@/views/NewListView'
-import PostView from '@/views/PostView'
-import Login from '@/views/Login'
-import Register from '@/views/Register'
+import PostDetailsView from '@/views/PostDetailsView'
+import LoginView from '@/views/LoginView'
+import RegisterView from '@/views/RegisterView'
 import RecomendationListView from '@/views/RecomendationListView'
 
 Vue.use(Router)
@@ -19,8 +19,8 @@ export default new Router({
     { path: '/top/:page(\\d+)?', component: TopListView },
     { path: '/new/:page(\\d+)?', component: NewListView },
     { path: '/recommendations/:page(\\d+)?', component: RecomendationListView },
-    { path: '/post/:id([A-Za-z0-9-_]+)?', component: PostView },
-    { path: '/login', component: Login },
-    { path: '/regsiter', component: Register }
+    { path: '/post/:id([A-Za-z0-9-_]+)?', component: PostDetailsView },
+    { path: '/login', component: LoginView },
+    { path: '/regsiter', component: RegisterView }
   ]
 })
